Add files via upload
Currently these are the low level files, just so that anyone interested in c++ coding can hopefully grasp a small concept of how the coding styles work for operating a .cpp.  I will be updating these as I go, with a single player build, and an online build.  Yes I am running these as 1 build, with optional configurations for the engine as follows: single-player, or multi-player, with support for JRPG, ARPG, Adventure, or just creating an online hack and slash game.  Currently deciding on coding with openGL, or Alegro 5. (this is a decision that has been a rough passage).   However I am uploading the code as strictly a text based through development so that you may choose the API, and client server protocols you would like (windows only dx12, winSOCK, or cross platform, ASIO, and openGL/allegro5)